feat: 로직 분리 및 jsdoc추가가
