+++ src/features/indicator/Indicator.types.ts
@@ -1,0 +1,5 @@
+export interface IndicatorProps {
+  totalPages: number;
+  currentPage: number;
+  onClick?: (index: number) => void;
+}
